requirements.txt
Pygments~=2.19.2
h11~=0.16.0
pillow~=11.3.0
pip~=25.2
typing_extensions~=4.15.0
distro~=1.9.0
PySocks~=1.7.1
cffi~=1.17.1
setuptools~=78.1.1
rich~=14.1.0
markdown-it-py~=4.0.0
toml~=0.10.2
requests~=2.32.5
colorama~=0.4.6
PyYAML~=6.0.3
zipp~=3.23.0
sniffio~=1.3.0
httpcore~=1.0.9
idna~=3.7
httpx~=0.28.1
jiter~=0.6.1
mdurl~=0.1.2
wheel~=0.45.1
openai~=1.109.1
pydantic~=2.11.9
websockets~=15.0.1
anyio~=4.10.0
pydantic_core~=2.33.2
psutil~=7.1.0
qrcode~=8.2
certifi~=2025.8.3
ncatbot=4.2.2.post4
packaging~=25.0
tqdm~=4.67.1
schedule~=1.2.2
aiofiles~=24.1.0
brotlicffi~=1.0.9.2
urllib3~=2.5.0
annotated-types~=0.6.0
charset-normalizer~=3.3.2
pycparser~=2.23